// src/app/modules/organizations/shared/enums/action-types.ts
export enum ActionTypes {
  GetOrganization = '[Organization] Get Organization',
  GetOrganizationSuccess = '[Organization] Get Organization Success',
  GetOrganizationError = '[Organization] Get Organization Error',

  GetOrganizations = '[Organization] Get Organizations',
  GetOrganizationsSuccess = '[Organization] Get Organizations Success',
  GetOrganizationsError = '[Organization] Get Organizations Error',

  PostOrganization = '[Organization] Add Organization',
  PostOrganizationSuccess = '[Organization] Add Organization Success',
  PostOrganizationError = '[Organization] Add Organization Error',

  UpdateOrganization = '[Organization] Update Organization',
  UpdateOrganizationSuccess = '[Organization] Update Organization Success',
  UpdateOrganizationError = '[Organization] Update Organization Error',

  DeleteOrganization = '[Organization] Delete Organization  ',
  DeleteOrganizationSuccess = '[Organization]  Delete Organization  Success',
  DeleteOrganizationError = '[Organization]  Delete Organization  Error',
}